Complete view/route restructure using the Flask Blueprint class for organization and easier changes
