+++ .gitignore
@@ -1,3 +1,3 @@
 testing/
 static/files/
-apikeys.txt
+routes/helpers/apikeys.txt
