--- apikeys.txt
@@ -1,3 +1,0 @@
-test1
-test2
-test3
