+++ routes/helpers/data_proc.py
@@ -1,5 +1,5 @@
 import os
-import utils
+from routes.helpers import utils
 import logging
 import pandas as pd
